+++ kernel/sched.c
@@ -964,64 +964,6 @@ void z_sched_init(void)
 #endif /* CONFIG_SCHED_CPU_MASK_PIN_ONLY */
 }
 
-#ifdef CONFIG_SCHED_DEADLINE
-void z_impl_k_thread_absolute_deadline_set(k_tid_t tid, int deadline)
-{
-	struct k_thread *thread = tid;
-
-	/* The prio_deadline field changes the sorting order, so can't
-	 * change it while the thread is in the run queue (dlists
-	 * actually are benign as long as we requeue it before we
-	 * release the lock, but an rbtree will blow up if we break
-	 * sorting!)
-	 */
-	K_SPINLOCK(&_sched_spinlock) {
-		if (z_is_thread_queued(thread)) {
-			dequeue_thread(thread);
-			thread->base.prio_deadline = deadline;
-			queue_thread(thread);
-		} else {
-			thread->base.prio_deadline = deadline;
-		}
-	}
-}
-
-void z_impl_k_thread_deadline_set(k_tid_t tid, int deadline)
-{
-
-	deadline = clamp(deadline, 0, INT_MAX);
-
-	int32_t newdl = k_cycle_get_32() + deadline;
-
-	z_impl_k_thread_absolute_deadline_set(tid, newdl);
-}
-
-#ifdef CONFIG_USERSPACE
-static inline void z_vrfy_k_thread_absolute_deadline_set(k_tid_t tid, int deadline)
-{
-	struct k_thread *thread = tid;
-
-	K_OOPS(K_SYSCALL_OBJ(thread, K_OBJ_THREAD));
-
-	z_impl_k_thread_absolute_deadline_set((k_tid_t)thread, deadline);
-}
-#include <zephyr/syscalls/k_thread_absolute_deadline_set_mrsh.c>
-
-static inline void z_vrfy_k_thread_deadline_set(k_tid_t tid, int deadline)
-{
-	struct k_thread *thread = tid;
-
-	K_OOPS(K_SYSCALL_OBJ(thread, K_OBJ_THREAD));
-	K_OOPS(K_SYSCALL_VERIFY_MSG(deadline > 0,
-				    "invalid thread deadline %d",
-				    (int)deadline));
-
-	z_impl_k_thread_deadline_set((k_tid_t)thread, deadline);
-}
-#include <zephyr/syscalls/k_thread_deadline_set_mrsh.c>
-#endif /* CONFIG_USERSPACE */
-#endif /* CONFIG_SCHED_DEADLINE */
-
 void z_impl_k_reschedule(void)
 {
 	k_spinlock_key_t key;
